backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]
cryptography
python-multipart
pydantic[email]
PyJWT
